Extract repair_job_dispatch.rs from turn.rs (parent #680)
Ninth vertical-slice extraction: move the verifier-repair state-machine
dispatch out of turn.rs into a new
src/agent/loop_run/repair_job_dispatch.rs sibling module. Largest single
extraction since the test-mod migrations.

Migrated 14 methods (all free fns over &mut Agent):

Entry points (pub(super)):
- dispatch_repair_job_step — RepairStep state-machine dispatch.
- dispatch_missing_verifier_job_step — missing-verifier bootstrap.
- repair_rejection_next_action — pure-fn error → next-action note.

Private branch handlers:
- handle_repair_job_verifier_pass / _failure / _no_verifier / _safe_stop
  (TaskContractVerifierOutcome branches)
- handle_repair_job_diagnostic_step / _patch_provider_step
  (RepairStep branches)
- drive_repair_job_verifier — verifier rerun + outcome routing.
- dispatch_after_repair_patch_rejection — rejected-patch routing.
- write_repair_job_step_status — iteration-status renderer.
- repair_job_diagnostic_skipped_error / repair_job_safe_stop_outcome —
  terminal outcome builders.

Visibility changes in turn.rs:
- emit_safe_stop_report_for_repair_terminal → pub(super) (called from
  the new module).

No facade re-export (DR3-001). Caller updates:
- actor_loop_flow.rs: agent.dispatch_repair_job_step(...) /
  agent.dispatch_missing_verifier_job_step(...) → free-fn form.
- turn_tests.rs: super::Agent::repair_rejection_next_action(...) →
  super::super::repair_job_dispatch::repair_rejection_next_action(...).
- repair_runner_contract_tests.rs::repair_job_run_verifier_dispatch_uses_
  job_preserving_path: include_str!("turn.rs") →
  include_str!("repair_job_dispatch.rs"); markers updated for free-fn
  signature; body assertion updated to drive_repair_job_verifier(agent,
  args).

turn.rs: 5,366 -> 4,860 (-506 LOC)
Cumulative from 39,489: -34,629 LOC (-87.7%)

Verification:
- cargo fmt --check
- cargo clippy --lib --all-targets -- -D warnings
- cargo test --lib (3,114 passed)
