+++ src/agent/loop_run/turn.rs
@@ -2,10 +2,7 @@ use super::active_job_arbiter::{
     LoopControlAction, LoopControlInputs, RecoveryOwner, build_active_job_selected_payload,
     determine_loop_control_action,
 };
-use super::actor_loop_flow::{
-    TaskContractVerifierFlowOutcome, format_iteration_status, repair_job_done_outcome,
-    run_actor_loop,
-};
+use super::actor_loop_flow::{format_iteration_status, run_actor_loop};
 use super::auto_test::{AutoTestKind, AutoTestRunner};
 use super::completion_evidence::is_repo_edit_no_op;
 use super::interrupt::{InterruptEnv, InterruptMonitor};
@@ -13,7 +10,6 @@ use super::repair_driver::{
     VERIFIER_REPAIR_PASS_WALL_CLOCK_LIMIT_SECS, VerifierRepairPassOutcome,
     verifier_repair_pass_timeout_error,
 };
-use super::repair_job::verifier_repair_context_from_failure;
 use super::repair_patch_validation::{
     CheapCheckOutcome, RepairRejectionSignal, ValidationFailure,
     build_verifier_repair_pass_ledger_outcome, validate_accepted_repair_plan_authorizes_target,
@@ -69,30 +65,26 @@ use super::tool_policy::{
 use super::verifier_diagnostic_attempt::{
     VERIFIER_DIAGNOSTIC_ATTEMPT_LIMIT, verifier_diagnostic_attempt_spec,
 };
-use super::verifier_driver::TaskContractVerifierOutcome;
 use super::verifier_orchestration::{
     JobInstallOutcome, PreparedVerifierDiagnosticPass, PreparedVerifierRepairPass,
-    TaskContractVerifierFlowArgs, VerifierDiagnosticPassOutcome, VerifierRepairAttemptProgress,
+    VerifierDiagnosticPassOutcome, VerifierRepairAttemptProgress,
     build_verifier_exit_zero_evidence, emit_patch_proposal_legacy_validation_comparison_event,
-    emit_patch_proposal_shadow_validation_event, emit_repair_progress_classified_event,
-    repair_terminal_exit_reason, synthesized_missing_implementation_target_path_for_request,
-    synthesized_missing_test_target_path_for_request, task_contract_needs_verification,
-    task_contract_no_verifier_note, task_contract_verifier_failure_attempt_limit,
-    task_contract_verifier_repair_note, task_contract_verifier_targeted_edit_required_note,
-    test_target_path_compatible_with_request, validate_verifier_repair_intents_with_accepted_plan,
-    verifier_diagnostic_messages, verifier_repair_context_diagnostics,
-    verifier_repair_diagnostic_pending_note, verifier_repair_intent_limits,
-    verifier_repair_pass_messages, verifier_repair_pass_request_error_message,
-    verifier_repair_policy_for_target_hint, verifier_repair_safe_stop_message,
-    verifier_repair_target_display, verifier_repair_transition_message,
-    verifier_repair_unsafe_target_message, verifier_setup_policy_message,
+    emit_patch_proposal_shadow_validation_event,
+    synthesized_missing_implementation_target_path_for_request,
+    synthesized_missing_test_target_path_for_request, task_contract_no_verifier_note,
+    task_contract_verifier_targeted_edit_required_note, test_target_path_compatible_with_request,
+    validate_verifier_repair_intents_with_accepted_plan, verifier_diagnostic_messages,
+    verifier_repair_context_diagnostics, verifier_repair_diagnostic_pending_note,
+    verifier_repair_intent_limits, verifier_repair_pass_messages,
+    verifier_repair_pass_request_error_message, verifier_repair_policy_for_target_hint,
+    verifier_repair_safe_stop_message, verifier_repair_target_display,
+    verifier_repair_transition_message, verifier_repair_unsafe_target_message,
+    verifier_setup_policy_message,
 };
 use super::verifier_repair_shadow::legacy_repair_brief_input_from_assessment;
-use super::verifier_repair_targeting::changed_files_for_verifier;
 use super::workspace_candidates::existing_workspace_candidate_for_role_in_scope;
 use super::workspace_walk::workspace_appears_empty;
 use super::*;
-use crate::agent::orchestration::verify_repo_progress;
 use crate::logging::{log_llm_event, stable_path_hash};
 use crate::model_capabilities::model_capabilities;
 use crate::modes::plan_act::WorkMode;
@@ -637,512 +629,6 @@ impl Agent {
         }
     }
 
-    fn handle_repair_job_verifier_pass(
-        &mut self,
-        args: TaskContractVerifierFlowArgs<'_, '_>,
-        previous_repair_context: super::repair_job::RepairJob,
-        command: String,
-    ) -> TaskContractVerifierFlowOutcome {
-        if task_contract_needs_verification(
-            self.session.mode_state.mode,
-            args.task_contract,
-            &self.task_contract_evidence_set_this_turn,
-        ) {
-            return TaskContractVerifierFlowOutcome::Exit {
-                reason: ExitReason::MissingVerification,
-                error_text: "verifier passed but verifier evidence could not be recorded"
-                    .to_string(),
-            };
-        }
-        let safe_command = crate::session::feedback::mask_secrets(&command).replace('`', "\\`");
-        if let Some(sanitized) =
-            super::verifier_skill::sanitize_verify_command_for_case_record(&command)
-        {
-            args.task_contract_verify_commands_collected.push(sanitized);
-        }
-        if let Some(job) = self.repair_job.as_mut() {
-            job.apply_event(super::repair_job::RepairJobEvent::VerifierObserved {
-                delta: super::repair_job::VerifierDelta::Passed,
-            });
-        }
-        emit_repair_progress_classified_event(
-            self.session_store.session_id(),
-            Some(&previous_repair_context),
-            self.repair_job.as_ref(),
-            true,
-        );
-        self.missing_verifier_job = None;
-        *args.verifier_repair_retries = 0;
-        self.repair_job_artifact_attempts = 0;
-        *args.task_contract_verifier_passed_in_loop = true;
-        self.task_contract_verifier_passed_this_actor_loop = true;
-        TaskContractVerifierFlowOutcome::Done {
-            final_prose: format!(
-                "Completed requested repository changes and verified them with `{safe_command}`."
-            ),
-        }
-    }
-
-    fn handle_repair_job_verifier_failure(
-        &mut self,
-        args: TaskContractVerifierFlowArgs<'_, '_>,
-        previous_repair_context: super::repair_job::RepairJob,
-        changed_files: &[String],
-        command: String,
-        output: String,
-    ) -> TaskContractVerifierFlowOutcome {
-        *args.contract_verification_retries += 1;
-        let mut repair_context = verifier_repair_context_from_failure(
-            &self.work_root,
-            &command,
-            &output,
-            changed_files,
-            *args.contract_verification_retries,
-            Some(&previous_repair_context),
-        );
-        let attempt_limit =
-            task_contract_verifier_failure_attempt_limit(Some(&previous_repair_context));
-        if *args.contract_verification_retries >= attempt_limit {
-            self.repair_job = Some(repair_context);
-            super::verifier_orchestration::emit_safe_stop_report_for_repair_exhausted(self);
-            return TaskContractVerifierFlowOutcome::Exit {
-                reason: ExitReason::RepairExhausted,
-                error_text: format!(
-                    "verifier repair budget exhausted: {}\n{}",
-                    crate::session::feedback::mask_secrets(&command),
-                    crate::session::feedback::mask_secrets(&output)
-                ),
-            };
-        }
-
-        let applied_outcome_promotion = super::repair_job::apply_verifier_rerun_observation(
-            &mut repair_context,
-            Some(&previous_repair_context),
-        );
-        *args.contract_verifier_repair_edit_count = Some(args.repo_edit_calls_made_this_turn);
-        self.task_contract_verifier_repair_pending = true;
-        if let Some(outcome) = repair_context.rerun_outcome {
-            log_llm_event(
-                "agent.verifier_repair.rerun_classified",
-                serde_json::json!({
-                    "session_id": self.session_store.session_id(),
-                    "outcome": outcome.as_str(),
-                    "previous_failure_signature": repair_context.previous_failure_signature.as_deref(),
-                    "current_failure_signature": repair_context.failure_signature.as_str(),
-                    "previous_failure_count": repair_context.previous_failure_count,
-                    "current_failure_count": repair_context.failure_count,
-                    "job_preserving": true,
-                }),
-            );
-        }
-        emit_repair_progress_classified_event(
-            self.session_store.session_id(),
-            Some(&previous_repair_context),
-            Some(&repair_context),
-            false,
-        );
-        self.repair_job = Some(repair_context);
-        super::verifier_orchestration::maybe_emit_repair_exhausted_from_promotion(
-            self,
-            applied_outcome_promotion,
-        );
-        *args.repo_change_retries = 0;
-        *args.verifier_repair_retries = 0;
-        self.repair_job_artifact_attempts = 0;
-        write_stdout_rendered(
-            &format_iteration_status(
-                args.last_iter,
-                self.config.max_iterations,
-                "Verification failed",
-                "Repair job observed verifier failure and updated its state.",
-                self.footer.current_cols(),
-            ),
-            true,
-        );
-        self.push_system_note(task_contract_verifier_repair_note(
-            &command,
-            &output,
-            *args.contract_verification_retries,
-            attempt_limit,
-            self.repair_job.as_ref(),
-        ));
-        TaskContractVerifierFlowOutcome::Continue
-    }
-
-    fn handle_repair_job_verifier_no_verifier(
-        &mut self,
-        args: TaskContractVerifierFlowArgs<'_, '_>,
-    ) -> TaskContractVerifierFlowOutcome {
-        if let Some(job) = self.repair_job.as_mut() {
-            job.apply_event(super::repair_job::RepairJobEvent::VerifierObserved {
-                delta: super::repair_job::VerifierDelta::VerifierUnavailable,
-            });
-        }
-        self.task_contract_verifier_repair_pending = true;
-        *args.repo_change_retries = 0;
-        *args.verifier_repair_retries = 0;
-        write_stdout_rendered(
-            &format_iteration_status(
-                args.last_iter,
-                self.config.max_iterations,
-                "Verification missing",
-                "Repair job observed that no runnable verifier is available.",
-                self.footer.current_cols(),
-            ),
-            true,
-        );
-        TaskContractVerifierFlowOutcome::Continue
-    }
-
-    fn handle_repair_job_verifier_safe_stop(
-        &mut self,
-        last_iter: usize,
-        reason: super::task_contract::SafeStopReason,
-    ) -> TaskContractVerifierFlowOutcome {
-        let outcome = super::verifier_orchestration::handle_task_contract_verifier_safe_stop(
-            self,
-            last_iter,
-            reason,
-            "repair_job_verifier",
-        );
-        if let Some(job) = self.repair_job.as_mut() {
-            job.apply_event(super::repair_job::RepairJobEvent::VerifierObserved {
-                delta: super::repair_job::VerifierDelta::VerifierUnavailable,
-            });
-        }
-        outcome
-    }
-
-    pub(super) fn dispatch_repair_job_step(
-        &mut self,
-        args: TaskContractVerifierFlowArgs<'_, '_>,
-        repo_edit_calls_made_this_turn: &mut usize,
-    ) -> TaskContractVerifierFlowOutcome {
-        let Some(step) = self
-            .repair_job
-            .as_mut()
-            .map(|job| job.begin_next_repair_step())
-        else {
-            return TaskContractVerifierFlowOutcome::Continue;
-        };
-
-        match step {
-            super::repair_job::RepairStep::RunDiagnostic => {
-                self.handle_repair_job_diagnostic_step(args)
-            }
-            super::repair_job::RepairStep::RunPatchProvider { target_hint } => self
-                .handle_repair_job_patch_provider_step(
-                    args,
-                    repo_edit_calls_made_this_turn,
-                    target_hint,
-                ),
-            super::repair_job::RepairStep::RunVerifier => self.drive_repair_job_verifier(args),
-            super::repair_job::RepairStep::SafeStop { reason } => {
-                self.repair_job_safe_stop_outcome(reason)
-            }
-            super::repair_job::RepairStep::Done => repair_job_done_outcome(),
-        }
-    }
-
-    fn handle_repair_job_diagnostic_step(
-        &mut self,
-        args: TaskContractVerifierFlowArgs<'_, '_>,
-    ) -> TaskContractVerifierFlowOutcome {
-        self.write_repair_job_step_status(
-            args.last_iter,
-            "Verifier diagnostic",
-            "Running short-lived diagnostic LLM pass outside the main session.",
-        );
-        match super::verifier_orchestration::run_verifier_diagnostic_pass(self) {
-            VerifierDiagnosticPassOutcome::Accepted => {
-                self.write_repair_job_step_status(
-                    args.last_iter,
-                    "Verifier diagnostic",
-                    "Accepted validated diagnostic result; continuing verifier repair.",
-                );
-                TaskContractVerifierFlowOutcome::Continue
-            }
-            VerifierDiagnosticPassOutcome::RetryPending { error } => {
-                self.write_repair_job_step_status(
-                    args.last_iter,
-                    "Verifier diagnostic",
-                    &format!("Diagnostic pass failed ({error}); retrying with fallback model."),
-                );
-                TaskContractVerifierFlowOutcome::Continue
-            }
-            VerifierDiagnosticPassOutcome::Unavailable { error } => {
-                self.emit_safe_stop_report_for_repair_terminal(
-                    super::repair_job::RepairTerminalReason::DiagnosticUnavailable,
-                );
-                TaskContractVerifierFlowOutcome::Exit {
-                    reason: ExitReason::VerifierFailed,
-                    error_text: format!("verifier repair diagnostic_unavailable: {error}"),
-                }
-            }
-            VerifierDiagnosticPassOutcome::Skipped => TaskContractVerifierFlowOutcome::Exit {
-                reason: ExitReason::VerifierFailed,
-                error_text: self.repair_job_diagnostic_skipped_error(),
-            },
-        }
-    }
-
-    fn handle_repair_job_patch_provider_step(
-        &mut self,
-        args: TaskContractVerifierFlowArgs<'_, '_>,
-        repo_edit_calls_made_this_turn: &mut usize,
-        target_hint: super::task_contract::RecoveryTargetHint,
-    ) -> TaskContractVerifierFlowOutcome {
-        self.write_repair_job_step_status(
-            args.last_iter,
-            "Verifier repair",
-            "Running controller-applied repair pass for the selected target.",
-        );
-        match super::verifier_orchestration::run_verifier_repair_pass_and_apply(self, &target_hint)
-        {
-            VerifierRepairPassOutcome::Applied { relative_path } => {
-                *repo_edit_calls_made_this_turn = repo_edit_calls_made_this_turn.saturating_add(1);
-                *args.repo_change_retries = 0;
-                *args.verifier_repair_retries = 0;
-                self.write_repair_job_step_status(
-                    args.last_iter,
-                    "Verifier repair",
-                    &format!(
-                        "Applied controller repair edit to {relative_path}; verifier will rerun."
-                    ),
-                );
-                TaskContractVerifierFlowOutcome::Continue
-            }
-            VerifierRepairPassOutcome::Invalid {
-                error,
-                repair_attempt_outcome,
-            } => {
-                super::verifier_orchestration::record_controller_verifier_repair_invalid(
-                    self,
-                    &error,
-                    repair_attempt_outcome,
-                );
-                self.dispatch_after_repair_patch_rejection(args.last_iter, error, Some(target_hint))
-            }
-            VerifierRepairPassOutcome::Unavailable { relative_path } => {
-                super::verifier_orchestration::record_controller_verifier_repair_invalid(
-                    self,
-                    &format!(
-                        "verifier repair unavailable: no safe cheap check available for {relative_path}"
-                    ),
-                    None,
-                );
-                self.write_repair_job_step_status(
-                    args.last_iter,
-                    "Verifier repair",
-                    &format!(
-                        "No safe cheap check available for {relative_path}; continuing through repair job state."
-                    ),
-                );
-                TaskContractVerifierFlowOutcome::Continue
-            }
-            VerifierRepairPassOutcome::Skipped => TaskContractVerifierFlowOutcome::Exit {
-                reason: ExitReason::VerifierFailed,
-                error_text: "verifier repair patch provider skipped after committed dispatch"
-                    .to_string(),
-            },
-        }
-    }
-
-    fn write_repair_job_step_status(&self, last_iter: usize, title: &str, message: &str) {
-        write_stdout_rendered(
-            &format_iteration_status(
-                last_iter,
-                self.config.max_iterations,
-                title,
-                message,
-                self.footer.current_cols(),
-            ),
-            true,
-        );
-    }
-
-    fn repair_job_diagnostic_skipped_error(&mut self) -> String {
-        self.emit_safe_stop_report_for_repair_terminal(
-            super::repair_job::RepairTerminalReason::DiagnosticUnavailable,
-        );
-        "verifier repair diagnostic skipped after committed dispatch".to_string()
-    }
-
-    fn repair_job_safe_stop_outcome(
-        &mut self,
-        reason: super::repair_job::RepairTerminalReason,
-    ) -> TaskContractVerifierFlowOutcome {
-        self.emit_safe_stop_report_for_repair_terminal(reason);
-        TaskContractVerifierFlowOutcome::Exit {
-            reason: repair_terminal_exit_reason(reason),
-            error_text: format!("verifier repair safe stop: {}", reason.as_str()),
-        }
-    }
-
-    fn drive_repair_job_verifier(
-        &mut self,
-        args: TaskContractVerifierFlowArgs<'_, '_>,
-    ) -> TaskContractVerifierFlowOutcome {
-        let Some(previous_repair_context) = self.repair_job.clone() else {
-            return TaskContractVerifierFlowOutcome::Continue;
-        };
-        let current_verif = verify_repo_progress(args.before_snapshot, &self.work_root);
-        let changed_files = changed_files_for_verifier(args.accumulated, &current_verif);
-        write_stdout_rendered(
-            &format_iteration_status(
-                args.last_iter,
-                self.config.max_iterations,
-                "Task contract",
-                "Running verifier for repair job result.",
-                self.footer.current_cols(),
-            ),
-            true,
-        );
-
-        match super::verifier_orchestration::run_task_contract_verifier_once(self, &changed_files) {
-            TaskContractVerifierOutcome::Passed { command } => {
-                self.handle_repair_job_verifier_pass(args, previous_repair_context, command)
-            }
-            TaskContractVerifierOutcome::Failed { command, output } => self
-                .handle_repair_job_verifier_failure(
-                    args,
-                    previous_repair_context,
-                    &changed_files,
-                    command,
-                    output,
-                ),
-            TaskContractVerifierOutcome::NoVerifier => {
-                self.handle_repair_job_verifier_no_verifier(args)
-            }
-            TaskContractVerifierOutcome::Disabled => TaskContractVerifierFlowOutcome::Exit {
-                reason: ExitReason::MissingVerification,
-                error_text: "task contract requires verification, but ANVIL_NO_AUTO_TEST is set"
-                    .to_string(),
-            },
-            TaskContractVerifierOutcome::TransportError { error } => {
-                TaskContractVerifierFlowOutcome::Exit {
-                    reason: ExitReason::TransportError,
-                    error_text: error,
-                }
-            }
-            TaskContractVerifierOutcome::SafeStop { reason } => {
-                self.handle_repair_job_verifier_safe_stop(args.last_iter, reason)
-            }
-        }
-    }
-
-    pub(super) fn dispatch_missing_verifier_job_step(
-        &mut self,
-        args: TaskContractVerifierFlowArgs<'_, '_>,
-        next_action: super::repair_job::VerifierBootstrapNextAction,
-    ) -> Option<TaskContractVerifierFlowOutcome> {
-        match next_action {
-            super::repair_job::VerifierBootstrapNextAction::RequestSetupEdit => None,
-            super::repair_job::VerifierBootstrapNextAction::RerunVerifier => Some(
-                super::verifier_orchestration::drive_task_contract_verifier(self, args),
-            ),
-            super::repair_job::VerifierBootstrapNextAction::SafeStop { reason } => {
-                self.emit_safe_stop_report_for_verifier_missing();
-                Some(TaskContractVerifierFlowOutcome::Exit {
-                    reason: ExitReason::MissingVerification,
-                    error_text: reason.to_string(),
-                })
-            }
-        }
-    }
-
-    fn dispatch_after_repair_patch_rejection(
-        &mut self,
-        last_iter: usize,
-        error: String,
-        attempted_target_hint: Option<super::task_contract::RecoveryTargetHint>,
-    ) -> TaskContractVerifierFlowOutcome {
-        let Some(action) = self.repair_job.as_ref().map(|job| job.next_action()) else {
-            return TaskContractVerifierFlowOutcome::Exit {
-                reason: ExitReason::VerifierFailed,
-                error_text: error,
-            };
-        };
-        match action {
-            super::repair_job::RepairNextAction::SafeStop { reason } => {
-                self.emit_safe_stop_report_for_repair_terminal(reason);
-                TaskContractVerifierFlowOutcome::Exit {
-                    reason: repair_terminal_exit_reason(reason),
-                    error_text: format!(
-                        "verifier repair safe stop after rejected patch: {error}. next_action: {}",
-                        Self::repair_rejection_next_action(&error)
-                    ),
-                }
-            }
-            super::repair_job::RepairNextAction::RequestDiagnostic
-            | super::repair_job::RepairNextAction::Replan => {
-                write_stdout_rendered(
-                    &format_iteration_status(
-                        last_iter,
-                        self.config.max_iterations,
-                        "Verifier repair",
-                        "Rejected invalid controller repair proposal; repair job will re-run diagnostics.",
-                        self.footer.current_cols(),
-                    ),
-                    true,
-                );
-                TaskContractVerifierFlowOutcome::Continue
-            }
-            super::repair_job::RepairNextAction::RequestPatch { target_hint } => {
-                let target_changed = attempted_target_hint
-                    .as_ref()
-                    .map(|attempted| attempted.path != target_hint.path)
-                    .unwrap_or(false);
-                let note = if target_changed {
-                    "Rejected invalid controller repair proposal; repair job selected another target."
-                } else {
-                    "Rejected invalid controller repair proposal; repair job will retry within its budget."
-                };
-                write_stdout_rendered(
-                    &format_iteration_status(
-                        last_iter,
-                        self.config.max_iterations,
-                        "Verifier repair",
-                        note,
-                        self.footer.current_cols(),
-                    ),
-                    true,
-                );
-                TaskContractVerifierFlowOutcome::Continue
-            }
-            super::repair_job::RepairNextAction::RerunVerifier => {
-                TaskContractVerifierFlowOutcome::Continue
-            }
-            super::repair_job::RepairNextAction::VerifiedDone => {
-                TaskContractVerifierFlowOutcome::Done {
-                    final_prose:
-                        "Completed requested repository changes and verified them with the required verifier."
-                            .to_string(),
-                }
-            }
-        }
-    }
-
-    pub(super) fn repair_rejection_next_action(error: &str) -> &'static str {
-        let normalized = error.to_ascii_lowercase();
-        if normalized.contains("role_mismatch") {
-            return "diagnostic and patch target disagreed; narrow the requested repair target or provide the authoritative file to change";
-        }
-        if normalized.contains("ambiguous")
-            || normalized.contains("authority")
-            || normalized.contains("expectation")
-        {
-            return "clarify the expected behavior or provide authoritative examples before retrying";
-        }
-        if normalized.contains("malformed") {
-            return "retry with a narrower task or simpler verifier output so the patch proposal can be structured safely";
-        }
-        if normalized.contains("duplicate") || normalized.contains("noop") {
-            return "inspect the verifier failure and retry with a different repair target";
-        }
-        "inspect the verifier diagnostics and retry with a narrower repair target"
-    }
-
     pub(super) fn record_missing_verifier_setup_failure(
         &mut self,
         last_iter: usize,
@@ -1176,7 +662,7 @@ impl Agent {
         false
     }
 
-    fn emit_safe_stop_report_for_repair_terminal(
+    pub(super) fn emit_safe_stop_report_for_repair_terminal(
         &mut self,
         reason: super::repair_job::RepairTerminalReason,
     ) {
